factor out roles helper into module
Becuase `Doorkeeper::ApplicationController` does not inherit from our `ApplicationController`, it does not get helper methods defined in `ApplicationController` ... but it can get mixins.
